requirements.txt
altair==5.5.0
attrs==25.4.0
azure-cognitiveservices-search-websearch==2.0.1
azure-common==1.1.28
azure-core==1.38.1
azure-mgmt-core==1.6.0
beautifulsoup4==4.14.3
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
gitdb==4.0.12
GitPython==3.1.46
google-search-results==2.4.2
idna==3.11
isodate==0.7.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
MarkupSafe==3.0.3
msrest==0.7.1
narwhals==2.16.0
numpy==2.0.2
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pillow==11.3.0
protobuf==6.33.5
pyarrow==21.0.0
pydeck==0.9.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.2
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
soupsieve==2.8.3
streamlit==1.50.0
tenacity==9.1.2
text-unidecode==1.3
toml==0.10.2
tornado==6.5.4
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
